requirements.txt
allure-pytest==2.13.5
greenlet==3.0.3
pytest==8.2.2
requests==2.32.3
selenium==4.22.0
SQLAlchemy==1.4.50
typing_extensions==4.12.2
